Removed unnecessary performance tests and renamed things to be sensible.
